021 - Final app
App will contain a simple h1 header with a name of app and a button. Text of button will change between ‘show details’ and ‘hide details’ (after each click). When button text will be ‘show detail’ next to button will be created a paragraph with a short string. If button text will be ‘hide details’ paragraph will be hided.